requirements.txt
scikit-learn
numpy
pandas
dvc
joblib
pyyaml
flask